Add texture sampler
A texture coord is read from obj for each vertex. The texture coords of pixels in the triangles (not vertex) are calculated using barycentric. For each pixel, the texture coord is calculated, then, the color value is read from texture according to this coord.

Degenerate triangles are removed in Barycentric function.